requirements.txt
dj-database-url==0.5.0
Django==2.2.17
djangorestframework==3.12.2
mysqlclient==2.0.1
python-decouple==3.3
pytz==2020.4
sqlparse==0.4.1
